refactor(game): inline BlackjackPolicy logic into Game class
Consolidated the BlackjackPolicy functionality directly into the Game
