+++ src/main/java/com/itacademy/blackjack/BlackjackApplication.java
@@ -2,7 +2,7 @@ package com.itacademy.blackjack;
 
 import com.itacademy.blackjack.deck.model.Deck;
 import com.itacademy.blackjack.deck.model.ScoringService;
-import com.itacademy.blackjack.game.domain.BlackjackPolicy;
+
 import com.itacademy.blackjack.game.model.Game;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -18,9 +18,8 @@ public class BlackjackApplication {
 		System.out.println(deck.draw());
 
 		ScoringService scoringService = new ScoringService();
-		BlackjackPolicy blackjackPolicy = new BlackjackPolicy();
 
-		Game game = new Game(scoringService,blackjackPolicy);
+		Game game = new Game(scoringService);
 		game.dealInitialCards();
 		System.out.println(game.getPlayerHand());
 		System.out.println(game.getCrupierHand());
